% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_hull.R
\name{in_hull}
\alias{in_hull}
\title{Find if a point is in the convex hull of a set of points}
\usage{
in_hull(points, inhull)
}
\arguments{
\item{points}{A set of 2D points. Accepts a dataframe of x & y coords, a numeric vector or a list of vectors.}

\item{inhull}{A list of vectors of the points to be checked.}
}
\value{
A vector of bools corresponding to if that point is in the convex hull or not.
}
\description{
Find if a point is in the convex hull of a set of points
}
\examples{
Points <- list(c(1,1),c(2,1),c(1,2),c(2,2),c(1.5,1.5),c(1.5,2.5))
check  <- list(c(1,1),c(1.5,1.5))
jarvis_march(Points, check)
}
