% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jarvis_march.R
\name{jarvis_march}
\alias{jarvis_march}
\title{Find the convex hull of a set of points in euclidean space}
\usage{
jarvis_march(point, plot = FALSE)
}
\arguments{
\item{point}{A set of 2D points. Accepts a dataframe of two columns of x & y coords, a numeric vector alternating x and y coords, or a list of vectors.}

\item{plot}{if TRUE the function outputs a plot of the convex hull.}
}
\value{
A dataframe consisting of four columns. Each row is a line segment of the convex hull and the columns are x and y coordinates.
}
\description{
Find the convex hull of a set of points in euclidean space
}
\examples{
Points <- list(c(1,1),c(2,1),c(1,2),c(2,2),c(1.5,1.5),c(1.5,2.5))
jarvis_march(Points, plot=TRUE)
}
