% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jarvis_march_plot.R
\name{jarvis_march_plot}
\alias{jarvis_march_plot}
\title{A function that returns the plot of the convex hull of a set of points}
\usage{
jarvis_march_plot(point)
}
\arguments{
\item{point}{A set of 2D points. Accepts a dataframe of two columns of x & y coords, a numeric vector alternating x and y coords, or a list of vectors.}
}
\value{
A ggplot2 object of the plot of the convex hull of the input set.
}
\description{
A function that returns the plot of the convex hull of a set of points
}
\examples{
Points <- list(c(1,1),c(2,1),c(1,2),c(2,2),c(1.5,1.5),c(1.5,2.5))
jarvis_march_plot(Points)
g <- jarvis_march_plot(Points)
}
